Create Roman_to_int.java
Roman to Integer solution 2.  In Right-to-Left pass, we can start from the second last index 
of the array, storing the integer value of the last character in 
some variable. We keep the integral value of the last character 
to the result beforehand. Now, as we move from right to left, we 
check at every step if the integer value of the current character is 
less than the last(previous/right) element we have seen. If it is less 
than the last value, we subtract this value from the result. Otherwise, 
we add it to our result. This implementation is totally based on the 
above fact that smaller-valued character before a bigger-valued one means 
it has to be subtracted from the latter one.